move formatting funcs
